protect edit and delete listing

+++ app/Http/Controllers/ListingController.php
@@ -58,6 +58,11 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing) {
+        // Ensure logged in user is owner
+        if($listing->user->id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -78,6 +83,11 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing) {
+        // Ensure logged in user is owner
+        if($listing->user->id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $listing->delete();
         return redirect("/")->with('message', 'Listing deleted!');
     }
